moving Customer_tab and Department_tab into ui folder
