+++ src/sample/ui/Customer_Tab/AddCustomer/AddCustomer.java
@@ -1,4 +1,4 @@
-package sample.Department_Tab.AddDepartment;
+package sample.ui.Customer_Tab.AddCustomer;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
@@ -6,12 +6,12 @@ import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
 
-public class AddDepartment extends Application {
+public class AddCustomer extends Application {
 
     @Override
     public void start(Stage primaryStage) throws Exception{
-        Parent root = FXMLLoader.load(getClass().getResource("AddDepartment.fxml"));
-        primaryStage.setTitle("Add Dept");
+        Parent root = FXMLLoader.load(getClass().getResource("AddCustomer.fxml"));
+        primaryStage.setTitle("Add Cus");
         primaryStage.setScene(new Scene(root, 600, 500));
         primaryStage.show();
     }
@@ -21,4 +21,3 @@ public class AddDepartment extends Application {
         launch(args);
     }
 }
-
