+++ src/sample/ui/Customer_Tab/ViewCustomers/ViewCustomers.java
@@ -1,4 +1,4 @@
-package sample.Department_Tab.DeleteDepartment;
+package sample.ui.Customer_Tab.ViewCustomers;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
@@ -6,12 +6,12 @@ import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
 
-public class DeleteDepartment extends Application {
+public class ViewCustomers extends Application {
 
     @Override
     public void start(Stage primaryStage) throws Exception{
-        Parent root = FXMLLoader.load(getClass().getResource("DeleteDepartment.fxml"));
-        primaryStage.setTitle("Delete Dept");
+        Parent root = FXMLLoader.load(getClass().getResource("ViewCustomers.fxml"));
+        primaryStage.setTitle("View Cus");
         primaryStage.setScene(new Scene(root, 600, 500));
         primaryStage.show();
     }
