+++ src/sample/ui/Department_Tab/AddDepartment/AddDepartmentController.java
@@ -1,4 +1,4 @@
-package sample.Department_Tab.AddDepartment;
+package sample.ui.Department_Tab.AddDepartment;
 
 import com.jfoenix.controls.JFXButton;
 import com.jfoenix.controls.JFXTextField;
