+++ src/sample/ui/Department_Tab/ModifyDepartment/ModifyDepartment.java
@@ -1,4 +1,4 @@
-package sample.Department_Tab.ViewDepartments;
+package sample.ui.Department_Tab.ModifyDepartment;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
@@ -6,12 +6,12 @@ import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
 
-public class ViewDepartments extends Application {
+public class ModifyDepartment extends Application{
 
     @Override
     public void start(Stage primaryStage) throws Exception{
-        Parent root = FXMLLoader.load(getClass().getResource("ViewDepartments.fxml"));
-        primaryStage.setTitle("View Dept");
+        Parent root = FXMLLoader.load(getClass().getResource("ModifyDepartment.fxml"));
+        primaryStage.setTitle("Modify Dept");
         primaryStage.setScene(new Scene(root, 600, 500));
         primaryStage.show();
     }
