+++ src/sample/ui/Department_Tab/ViewDepartments/ViewDepartmentsController.java
@@ -1,19 +1,16 @@
-package sample.Customer_Tab.ViewCustomers;
+package sample.ui.Department_Tab.ViewDepartments;
 
 import com.jfoenix.controls.JFXButton;
 import javafx.fxml.FXML;
 import javafx.scene.control.TableColumn;
 
-public class ViewCustomersController {
+public class ViewDepartmentsController {
 
     @FXML
     private TableColumn<?, ?> id_column;
 
     @FXML
     private TableColumn<?, ?> name_column;
-
-    @FXML
-    private TableColumn<?, ?> phone_column;
 
     @FXML
     private JFXButton cancle;
